Add full address bind configuration support to tgstation-server
This allows you to specify the configuration option BindAddress, in the format "127.0.0.1:5432", this string is then parsed by
System.Net.IPEndPoint.tryParse method, and turned into an instance of IPEndpoint available via the IServerAddressProvider interface

I've plugged in the EndPoint.Port value where appropriate into existing code, and made the endpoint get passed into kestrel's builder

By default it binds on 127.0.0.0:5000

The kestrel configuration is now fully deprecated and no longer supported, evidence it still exists will lead to a thrown configuration
error

In light of these breaking changes, the config version is bumped to
4.0.0

+++ src/Tgstation.Server.Host/Components/Interop/Bridge/RuntimeInformation.cs
@@ -22,7 +22,7 @@ namespace Tgstation.Server.Host.Components.Interop.Bridge
 		/// <summary>
 		/// The port the HTTP server is running on.
 		/// </summary>
-		public ushort ServerPort { get; }
+		public int ServerPort { get; }
 
 		/// <summary>
 		/// If DD should just respond if it's API is working and then exit.
@@ -65,7 +65,7 @@ namespace Tgstation.Server.Host.Components.Interop.Bridge
 			Version serverVersion,
 			string instanceName,
 			DreamDaemonSecurity? securityLevel,
-			ushort serverPort,
+			int serverPort,
 			bool apiValidateOnly)
 			: base(chatTrackingContext?.Channels ?? throw new ArgumentNullException(nameof(chatTrackingContext)))
 		{
